+++ src/main/java/com/allmappings/employeebackendapis/repository/EmployeeAddressRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployeeAddressRepository extends JpaRepository<EmployeeAddress, Integer> {
+public interface EmployeeAddressRepository extends JpaRepository<EmployeeAddress, Long> {
 }
