+++ src/main/java/com/allmappings/employeebackendapis/repository/EmployeeDepartmentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployeeDepartmentRepository extends JpaRepository<EmployeeDepartment, Integer> {
+public interface EmployeeDepartmentRepository extends JpaRepository<EmployeeDepartment, Long> {
 }
