+++ src/main/java/com/allmappings/employeebackendapis/repository/EmployeeSalaryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployeeSalaryRepository extends JpaRepository<EmployeeSalary, Integer> {
+public interface EmployeeSalaryRepository extends JpaRepository<EmployeeSalary, Long> {
 }
